@@ -12,8 +12,8 @@
   },
   "dependencies": {
     "firebase": "^10.12.2",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1"
   },
   "devDependencies": {
     "@types/node": "^20.14.2",
